Separate exception handling from point duplication
